fix: reconstruct the CallbackExample
